src/main/java/com/github/teamreflog/reflogserver/member/service/MemberService.java
package com.github.teamreflog.reflogserver.member.service;

import com.github.teamreflog.reflogserver.member.domain.MemberRepository;
import com.github.teamreflog.reflogserver.member.dto.MemberJoinRequest;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class MemberService {

    private final MemberRepository memberRepository;

    public void createMember(final MemberJoinRequest request) {
        memberRepository.save(request.toEntity());
    }
}
